moved auto commands into command folder
