+++ src/main/java/frc/robot/commands/AutoNamedCommands.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.commands;
 
 import com.pathplanner.lib.auto.NamedCommands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
